Move app template filters to a separate module
