+++ scout/server/app.py
@@ -2,23 +2,13 @@
 
 import logging
 import os
-import re
 from datetime import timedelta
-from typing import Dict, List, Optional, Union
-from urllib.parse import parse_qsl, unquote, urlsplit
 
 from flask import Flask, redirect, request, url_for
 from flask_cors import CORS
 from flask_login import current_user
-from markdown import markdown as python_markdown
-from markupsafe import Markup
 
 from scout import __version__
-from scout.constants import (
-    REVEL_SCORE_LABEL_COLOR_MAP,
-    SPIDEX_HUMAN,
-    SPLICEAI_SCORE_LABEL_COLOR_MAP,
-)
 from scout.log import init_log
 from scout.utils.config import load_config
 
@@ -43,6 +33,7 @@ from .blueprints import (
     variant,
     variants,
 )
+from .filters import register_filters as register_template_filters
 
 LOG = logging.getLogger(__name__)
 
@@ -83,7 +74,7 @@ def create_app(config_file=None, config=None):
     init_log(log=LOG, app=app)
     configure_extensions(app)
     register_blueprints(app)
-    register_filters(app)
+    register_template_filters(app)
     register_tests(app)
 
     if not (app.debug or app.testing) and app.config.get("MAIL_USERNAME"):
@@ -215,183 +206,6 @@ def register_blueprints(app):
     app.register_blueprint(institutes.overview)
     app.register_blueprint(managed_variants.managed_variants_bp)
     app.register_blueprint(omics_variants.omics_variants_bp)
-
-
-def register_filters(app):
-    """Creates methods that can be invoked from jinja2 or views/controllers, given that they are included app.custom_filters."""
-    app.custom_filters = type("CustomFilters", (), {})()  # Create an empty object as namespace
-
-    @app.template_filter()
-    def human_longint(value: Union[int, str]) -> str:
-        """Convert a long integers int or string representation into a human easily readable number."""
-        value = str(value)
-        if value.isnumeric():
-            return "{:,}".format(int(value)).replace(",", "&thinsp;")
-        return value
-
-    @app.template_filter()
-    def spidex_human(spidex):
-        """Translate SPIDEX annotation to human readable string."""
-        if spidex is None:
-            return "not_reported"
-        if abs(spidex) < SPIDEX_HUMAN["low"]["pos"][1]:
-            return "low"
-        if abs(spidex) < SPIDEX_HUMAN["medium"]["pos"][1]:
-            return "medium"
-        return "high"
-
-    @app.template_filter()
-    def get_label_or_color_by_score(
-        score: float,
-        map: str,
-        map_key: str,
-    ) -> str:
-        """Return a label or color for a given score based on predefined score ranges from the provided items_map."""
-        SCORE_ITEM_MAPS = {
-            "revel": REVEL_SCORE_LABEL_COLOR_MAP,
-            "spliceai": SPLICEAI_SCORE_LABEL_COLOR_MAP,
-        }
-        for (low, high), info in SCORE_ITEM_MAPS[map].items():
-            if low <= score <= high:
-                return info[map_key]
-
-    @app.template_filter()
-    def l2fc_2_fc(l2fc: float) -> float:
-        """Converts Log2 fold change to fold change."""
-        fc = 2 ** abs(l2fc)
-        return fc if l2fc >= 0 else -1 / fc
-
-    @app.template_filter()
-    def human_decimal(number, ndigits=4):
-        """Return a standard representation of a decimal number.
-
-        Args:
-            number (float): number to humanize
-            ndigits (int, optional): max number of digits to round to
-
-        Return:
-            str: humanized string of the decimal number
-        """
-        min_number = 10**-ndigits
-        if isinstance(number, str):
-            number = None
-        if number is None:
-            # NaN
-            return "-"
-        if number == 0:
-            # avoid confusion over what is rounded and what is actually 0
-            return 0
-        if number < min_number:
-            # make human readable and sane
-            return "<{}".format(min_number)
-
-        # round all other numbers
-        return round(number, ndigits)
-
-    @app.template_filter()
-    def markdown(text: str) -> Markup:
-        return Markup(python_markdown(text))
-
-    @app.template_filter()
-    def tuple_list_to_dict(tuple_list, key_elem, value_elem):
-        """Accepts a list of tuples and returns a dictionary with tuple element = key_elem as keys and tuple element = value_elem as values"""
-
-        return {tup[key_elem]: tup[value_elem] for tup in tuple_list}
-
-    @app.template_filter()
-    def url_decode(string):
-        """Decode a string with encoded hex values."""
-        return unquote(string)
-
-    @app.template_filter()
-    def url_args(url: str) -> Dict[str, str]:
-        """Return all arguments and values present in a string URL."""
-        return dict(parse_qsl(urlsplit(url).query))
-
-    @app.template_filter()
-    def cosmic_prefix(cosmicId):
-        """If cosmicId is an integer, add 'COSM' as prefix
-        otherwise return unchanged"""
-        if isinstance(cosmicId, int):
-            return "COSM" + str(cosmicId)
-        return cosmicId
-
-    @app.template_filter()
-    def format_variant_canonical_transcripts(variant: dict) -> List[str]:
-        lines = set()
-        genes = variant.get("genes") or []
-
-        for gene in genes:
-            gene_symbol = gene.get("hgnc_symbol", "")
-            hgvs = gene.get("hgvs_identifier") or ""
-            transcripts = gene.get("transcripts") or []
-
-            canonical_tx = None
-            primary_tx = None
-            tx_id = None
-
-            protein = None
-
-            for tx in transcripts:
-                tx_id = tx.get("transcript_id")
-                if not tx.get("is_canonical"):
-                    if tx.get("is_primary"):
-                        primary_tx = tx_id
-                    continue
-                canonical_tx = tx_id
-                protein = tx.get("protein_sequence_name") or ""
-                break
-
-            line_components = [f"{canonical_tx or primary_tx or tx_id} ({gene_symbol})"]
-            if hgvs:
-                line_components.append(unquote(hgvs))
-            if protein:
-                line_components.append(unquote(protein))
-
-            lines.add(" ".join(line_components))
-
-        return list(lines)
-
-    app.custom_filters.format_variant_canonical_transcripts = format_variant_canonical_transcripts
-
-    @app.template_filter()
-    def upper_na(string):
-        """
-        Uppercase ocurences of "dna" and "rna" for nice display.
-        """
-
-        return re.sub(r"[Dd][Nn][Aa]", r"DNA", re.sub(r"[Rr][Nn][Aa]", r"RNA", string))
-
-    @app.template_filter()
-    def count_cursor(pymongo_cursor):
-        """Count number of returned documents (deprecated pymongo.cursor.count())"""
-        # Perform operations on a copy of the cursor so original does not move
-        cursor_copy = pymongo_cursor.clone()
-        return len(list(cursor_copy))
-
-    @app.template_filter()
-    def list_intersect(list1, list2) -> list:
-        """Returns the elements that are common in 2 lists"""
-        return list(set(list1) & set(list2))
-
-    @app.template_filter()
-    def list_remove_none(in_list: list) -> list:
-        """Returns a list after removing any None values from it."""
-        return [item for item in in_list if item is not None]
-
-    @app.template_filter()
-    def spliceai_max(values) -> Optional[float]:
-        """Returns a list of SpliceAI values, extracting floats only from values like 'score:0.23'."""
-        float_values = []
-        for value in values:
-            if isinstance(value, str):
-                if ":" in value:  # Variant hits multiple genes
-                    value = value.split(":")[1].strip()
-            if value in [None, "-", "None"]:
-                continue
-            float_values.append(float(value))
-        if float_values:
-            return max(float_values)
 
 
 def register_tests(app):
